Move `ignore_changes` helper functions to a separate file.
